Move main.jsx and add new server dependencies
The main.jsx file has been moved into a Client subdirectory, and its import path for index.css has been modified. A new Server directory with a package.json and package-lock.json file has also been added. It comes with an array of dependencies required for the server functionality of the application.